requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
argcomplete==3.6.2
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.39.3
botocore==1.39.3
Brotli==1.1.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
choreographer==1.0.9
coloredlogs==15.0.1
contourpy==1.3.2
cuda-bindings==12.9.0
cuda-python==12.9.0
cycler==0.12.1
dacite==1.7.0
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
exceptiongroup==1.3.0
fiftyone==1.7.0
fiftyone-brain==0.21.2
fiftyone_db==1.3.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
gevent==25.5.1
geventhttpclient==2.3.4
glob2==0.7
grad-cam==1.5.5
graphql-core==3.2.6
greenlet==3.2.3
grpcio==1.67.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
humanize==4.12.3
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
inflate64==1.0.3
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonlines==4.0.0
kaleido==1.0.0
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
logistro==1.1.0
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multidict==6.6.3
multivolumefile==0.2.3
narwhals==1.45.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.17.0
onnxruntime==1.22.0
onnxslim==0.1.58
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pillow==11.3.0
plotly==6.2.0
pprintpp==0.4.0
priority==2.0.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
py7zr==1.0.0
pybcj==1.0.6
pycryptodomex==3.23.0
pydash==8.0.5
pymongo==4.9.2
pynndescent==0.5.13
pyparsing==3.2.3
pyppmd==1.2.0
python-dateutil==2.9.0.post0
python-rapidjson==1.20
pytorch-gradcam==0.2.1
pytz==2025.2
PyYAML==6.0.2
pyzstd==0.17.0
rarfile==4.2
regex==2024.11.6
requests==2.32.4
retrying==1.4.0
rtree==1.4.0
s3transfer==0.13.0
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.47.1
strawberry-graphql==0.275.5
sympy==1.14.0
tabulate==0.9.0
taskgroup==0.2.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
tritonclient==2.59.0
ttach==0.0.3
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.161
ultralytics-thop==2.0.14
umap-learn==0.5.9.post2
universal-analytics-python3==1.1.1
urllib3==2.5.0
voxel51-eta==0.14.2
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.20.1
zope.event==5.1
zope.interface==7.2
